Add basic instructions to generate and build Perl and Python bindings for eidlib.
These are not ready and need much more work. From the swig side nothing
is done do translate specific types or even wraping exceptions from c++
is done. Only thing done to the swig interface was add the specific
module name for each of the languages.

After compiling the simple_example in each of the examples folder should
work. The example just searches for the readers in the system, checks
if a card is inserted and dumps the basic public information in the card.

In each of the scripts the module path for our wrapper is hardcoded for 
"../GeneratedModule" because each language handles the search and
installation o modules different. Correctly installing these modules 
in a system will need more work.