Add Node.js wrappers generated with a git version of what will be swig3.
Swig repo with JS support - https://github.com/oliver----/swig-v8
Swig used compiled from commit e6dd124.

The way to compile nodejs modules is different from what we are used to,
it is needed to create a bindings.gyp file that will be run with the node-gyp
tool. 

Basic instructions are, install an updated nodejs version in the system, 
install node-gyp, install the nodejs headers with node-gyp. With this
the system is ready and now we change to eidlibNodeJS_Wrapper folder and run
node-gyp configure; node-gyp build.
Release version of the module is compiled in build/Release folder.
Check the example script to see how to use.

Basic example seems to work. With this, Python, Perl, PHP and Node.js wrappers
are at the same level of funcionality, meaning not ready for use before more
work is done.