Add checkings for the certificates status in the gui application.
At the start of the aplication request to load the certificates
so we can show the their status at the beginning in the main tab.

(Causes the application to be slower at start because all the cache is
getting created even when only wanting to check one cert.)

In the certificates tab, when selecting each cert always request
the getStatus() from the underground lib.

Comment some code that seems dead, remove next time.