Add laste details needed to fully compile and create msi with VS 2008.
Created new wix project file because changed some things there, adding 
the prebuild command the script to sign the minidriver. Also changed
the msi output folder.

With the changes both Product* files should now be possible to build 
both the VS2005 and VS2008 version. 
Maybe something is broken, but I hope not. They build the 32bits and 64bits
msi versions fully and signed.

Warning!:

I'm explicitly reverting for the moment, the changes aguerreiro did to the 
Product files to install the Openssl lib with another name.
I'm doing this because the projects versions I was migrating still had
the normal dlls defined for linkning etc and I wanted to finish this without
needing to yet again search for all the places to change it.
This revert is maked in the file with the text 
<!-- TODO: REFIX openssl path conflict with system -->.

More info in the wiki: http://projects.caixamagica.pt/cartaocidadao/wiki/CompilarMwVS2008