Implement support for expired cards - SOD validation changes
Due to a recent legislative action (DL 10-A/2020) related with the inpact of the COVID-19 pandemic
we need to support expired cards even when the SOD signing certificate is also expired.
This way the user can use all features of pteid Middleware - including PDF signatures that won't be considered
valid in standard PDF signature validators but the Portuguese legislation determines that indeed they
are valid if performed until the 30th of this year.