Change CA NID getters to return ASN1_OBJECT
On windows, openssl is statically linked to both applayer and cardlayer.
When openpace (only linked to cardlayer) creates its openssl objects,
these are only available on the object table from the cardlayer context.
These objects are needed on the applayer to find the NID used for chip
authentication; so the fix implemented is to directly return the
ASN1_OBJECTS from cardlayer instead of calling OBJ_nid2obj in the applayer.